fix(location): move target app editor dialog
